requirements.txt
langchain
openai
numpy
python-dotenv